+++ Modules/Intent.Modules.Entities/Templates/UpdateHelper/UpdateHelperTemplate.cs
@@ -22,7 +22,7 @@ namespace Intent.Modules.Entities.Templates.UpdateHelper
     /// Class to produce the template output
     /// </summary>
     
-    #line 1 "C:\Dev\Intent.Modules.NET\Modules\Intent.Modules.Entities\Templates\UpdateHelper\UpdateHelperTemplate.tt"
+    #line 1 "D:\Dev\Intent.Modules.NET\Modules\Intent.Modules.Entities\Templates\UpdateHelper\UpdateHelperTemplate.tt"
     [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.TextTemplating", "17.0.0.0")]
     public partial class UpdateHelperTemplate : CSharpTemplateBase<object>
     {
@@ -35,7 +35,7 @@ namespace Intent.Modules.Entities.Templates.UpdateHelper
             this.Write("using System;\r\nusing System.Collections.Generic;\r\n\r\n[assembly: DefaultIntentManag" +
                     "ed(Mode.Fully)]\r\n\r\nnamespace ");
             
-            #line 15 "C:\Dev\Intent.Modules.NET\Modules\Intent.Modules.Entities\Templates\UpdateHelper\UpdateHelperTemplate.tt"
+            #line 15 "D:\Dev\Intent.Modules.NET\Modules\Intent.Modules.Entities\Templates\UpdateHelper\UpdateHelperTemplate.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Namespace));
             
             #line default
@@ -43,7 +43,7 @@ namespace Intent.Modules.Entities.Templates.UpdateHelper
             this.Write("\r\n{\r\n    /// <summary>\r\n    /// Provides utility methods for updating collections" +
                     ".\r\n    /// </summary>\r\n    public static class ");
             
-            #line 20 "C:\Dev\Intent.Modules.NET\Modules\Intent.Modules.Entities\Templates\UpdateHelper\UpdateHelperTemplate.tt"
+            #line 20 "D:\Dev\Intent.Modules.NET\Modules\Intent.Modules.Entities\Templates\UpdateHelper\UpdateHelperTemplate.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(ClassName));
             
             #line default
@@ -60,25 +60,25 @@ namespace Intent.Modules.Entities.Templates.UpdateHelper
                     "signmentAction\">A delegate that defines how to update an item from the base coll" +
                     "ection using an item from the changed collection.</param>\r\n        /// <returns>" +
                     "The updated base collection.</returns>\r\n        /// <remarks>\r\n        /// If th" +
-                    "e changed collection is <see langword=\"null\" />, an empty list of type <see cref" +
-                    "=\"TOriginal\"/> will be returned.\r\n        /// If the base collection is <see lan" +
-                    "gword=\"null\" />, a new list of type <see cref=\"TOriginal\"/> will be created and " +
-                    "used.\r\n        /// </remarks>\r\n        public static ICollection<TOriginal> Crea" +
-                    "teOrUpdateCollection<TChanged, TOriginal>(\r\n            ICollection<TOriginal> b" +
-                    "aseCollection,\r\n            IEnumerable<TChanged>? changedCollection,\r\n         " +
-                    "   Func<TOriginal, TChanged, bool> equalityCheck,\r\n            Func<TOriginal?, " +
-                    "TChanged, TOriginal> assignmentAction)\r\n        {\r\n            if (changedCollec" +
-                    "tion == null)\r\n            {\r\n                return new List<TOriginal>();\r\n   " +
-                    "         }\r\n\r\n            baseCollection ??= new List<TOriginal>()!;\r\n\r\n        " +
-                    "    var result = baseCollection.CompareCollections(changedCollection, equalityCh" +
-                    "eck);\r\n            foreach (var elementToAdd in result.ToAdd)\r\n            {\r\n  " +
-                    "              var newEntity = assignmentAction(default, elementToAdd);\r\n        " +
-                    "        baseCollection.Add(newEntity);\r\n            }\r\n\r\n            foreach (va" +
-                    "r elementToRemove in result.ToRemove)\r\n            {\r\n                baseCollec" +
-                    "tion.Remove(elementToRemove);\r\n            }\r\n\r\n            foreach (var element" +
-                    "ToEdit in result.PossibleEdits)\r\n            {\r\n                assignmentAction" +
-                    "(elementToEdit.Original, elementToEdit.Changed);\r\n            }\r\n\r\n            r" +
-                    "eturn baseCollection;\r\n        }\r\n    }\r\n}");
+                    "e changed collection is <see langword=\"null\" />, an empty list of type <typepara" +
+                    "mref name=\"TOriginal\"/> will be returned.\r\n        /// If the base collection is" +
+                    " <see langword=\"null\" />, a new list of type <typeparamref name=\"TOriginal\"/> wi" +
+                    "ll be created and used.\r\n        /// </remarks>\r\n        public static ICollecti" +
+                    "on<TOriginal> CreateOrUpdateCollection<TChanged, TOriginal>(\r\n            IColle" +
+                    "ction<TOriginal> baseCollection,\r\n            IEnumerable<TChanged>? changedColl" +
+                    "ection,\r\n            Func<TOriginal, TChanged, bool> equalityCheck,\r\n           " +
+                    " Func<TOriginal?, TChanged, TOriginal> assignmentAction)\r\n        {\r\n           " +
+                    " if (changedCollection == null)\r\n            {\r\n                return new List<" +
+                    "TOriginal>();\r\n            }\r\n\r\n            baseCollection ??= new List<TOrigina" +
+                    "l>()!;\r\n\r\n            var result = baseCollection.CompareCollections(changedColl" +
+                    "ection, equalityCheck);\r\n            foreach (var elementToAdd in result.ToAdd)\r" +
+                    "\n            {\r\n                var newEntity = assignmentAction(default, elemen" +
+                    "tToAdd);\r\n                baseCollection.Add(newEntity);\r\n            }\r\n\r\n     " +
+                    "       foreach (var elementToRemove in result.ToRemove)\r\n            {\r\n        " +
+                    "        baseCollection.Remove(elementToRemove);\r\n            }\r\n\r\n            fo" +
+                    "reach (var elementToEdit in result.PossibleEdits)\r\n            {\r\n              " +
+                    "  assignmentAction(elementToEdit.Original, elementToEdit.Changed);\r\n            " +
+                    "}\r\n\r\n            return baseCollection;\r\n        }\r\n    }\r\n}");
             return this.GenerationEnvironment.ToString();
         }
     }